media_toolkit/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
beautifulsoup4==4.14.2
blinker==1.9.0
bs4==0.0.2
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
distro==1.9.0
Flask==3.1.2
google-api-core==2.25.1
google-auth==2.41.1
google-cloud-core==2.4.3
google-cloud-speech==2.33.0
google-cloud-storage==3.4.0
google-cloud-texttospeech==2.31.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.75.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
MarkupSafe==3.0.3
openai==2.0.0
proto-plus==1.26.1
protobuf==6.32.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.1.1
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
soupsieve==2.8
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
yt-dlp==2025.9.26
